+++ CHANGELOG.md
@@ -3,7 +3,7 @@ All notable changes to this project will be documented in this file.
 
 ## 0.4.4 - 2017-05-28
 ### Added
-- trapezoidal rule integration method
+- trapezoidal rule, simpson integration methods
 
 ## 0.4.3 - 2017-05-26
 ### Added