+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+## 0.4.5 - 2017-05-29
+### Added
+- exercise 7.3
+
 ## 0.4.4 - 2017-05-28
 ### Added
 - trapezoidal rule, simpson, romberg integration methods